# docs/openapi-banco-orbital.yaml
...
openapi: 3.0.3
info:
  title: Banco Orbital — API
  version: 1.0.0-draft
  description: |
    API del home banking de **Banco Orbital**. Este documento es el contrato entre
    nuestro backend y nuestro frontend, y la referencia para los otros bancos del
    curso que quieran comparar.

    ## Alcance de la v1

    Cuentas, tarjetas, préstamos y plazos fijos. **Seguros, CEDEARs y reservas
    quedan para la v2** — decisión tomada para que el alcance entre en el tiempo
    disponible.

    Los proveedores externos simulados (recargas de celular y empresas de
    servicios) **no están acá**: viven en una API aparte, `banco-proveedores`, que
    este banco consume por HTTP como si fuera un tercero real.

    ## Convenciones

    | Tema | Convención |
    |---|---|
    | Autenticación | JWT de Clerk en `Authorization: Bearer <token>` |
    | Errores | `{ "error": "mensaje" }`, con `details` opcional. Igual que el Banco Central |
    | Paginación | `page` y `limit` por query; respuesta `{ page, limit, count, data }` |
    | Importes | `number` con 2 decimales. Internamente se calculan en decimal exacto |
    | Nombres | Español, según `docs/GLOSARIO.md`. Campos en `snake_case`, como ya los devuelven los DTOs |
    | Idempotencia | Header `Idempotency-Key` (UUID v4) en **todo endpoint que mueve plata** |

    ### Sobre `monto` contra `importe`

    Dentro de nuestro banco el campo siempre se llama **`monto`**. `importe` aparece
...
    contrato. No mezclarlos.

    ### Sobre las rutas
...
servers:
...
security:
  - ClerkJWT: []
...
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de sesión emitido por Clerk. El backend lo valida y resuelve el
        `usuario` y su `persona` asociada. Todos los endpoints bajo `/api` lo
        requieren, además de usuario activo y perfil completo.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
      description: Página, empezando en 1.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Cantidad de elementos por página.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        UUID v4. Protege contra el doble envío: reintentar con la misma clave
        devuelve la respuesta original en vez de ejecutar la operación otra vez.
      example: 3f2504e0-4f89-41d3-9a0c-0305e82c3301
  responses:
    NoAutorizado:
      description: Falta el token, es inválido, o el usuario está inactivo.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No autorizado.' }
    Prohibido:
      description: El usuario no tiene permiso, o el recurso pertenece a otra persona.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No puedes operar sobre una cuenta que no te pertenece.' }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No existe el recurso solicitado.' }
    ErrorValidacion:
      description: El cuerpo o los parámetros no pasaron la validación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: 'Datos inválidos.'
            details:
              fieldErrors:
                monto: ['El monto debe ser mayor a cero.']
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje legible. En producción los 5xx se reemplazan por uno genérico.
          example: 'Saldo insuficiente.'
        details:
          type: object
          nullable: true
          description: |
            Detalle adicional. Presente en errores de validación (con `fieldErrors`)
            y en desarrollo. Nunca expone internos de la base en producción.
      additionalProperties: false
    Paginado:
      type: object
      description: Envoltorio de todos los listados paginados.
      required: [page, limit, count, data]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        count:
          type: integer
          description: Cantidad de elementos en **esta** página, no el total del recurso.
          example: 20
        data:
          type: array
          items: { type: object }
    Moneda:
      type: string
      enum: [ARS, USD]
      description: Moneda de la cuenta. La caja en pesos nace con la persona; la de dólares se abre aparte.
      example: ARS
    Monto:
      type: number
      format: double
      description: |
        Importe con 2 decimales. Se almacena como `NUMERIC(18,2)` y se calcula en
        decimal exacto; se expone como número por comodidad del cliente, que sólo
        lo formatea para mostrar.
      example: 184500.20
...
    PlazoFijo:
      type: object
      properties:
        id: { type: string, format: uuid }
        persona_id: { type: string, format: uuid }
        cuenta_id: { type: string, format: uuid }
        moneda: { $ref: '#/components/schemas/Moneda' }
        capital: { $ref: '#/components/schemas/Monto' }
        dias: { type: integer, example: 30 }
        tna:
          type: number
          description: TNA en porcentaje.
          example: 58
        interes: { $ref: '#/components/schemas/Monto' }
        total:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          description: '`capital + interes`. Es lo que se acredita al vencimiento.'
        tea:
          type: number
          description: Tasa Efectiva Anual equivalente, en porcentaje.
          example: 76.23
        fecha_constitucion: { type: string, format: date }
        fecha_vencimiento: { type: string, format: date }
        estado:
          type: string
          enum: [vigente, vencido, acreditado, cancelado_anticipado]
          example: vigente
paths:
...
  /api/plazos-fijos/simulaciones:
    post:
      tags: [Inversiones]
      summary: Simular un plazo fijo
      description: |
        **Método acordado — interés simple, base 365:**

        ```
        interes = capital × (tna / 100) × (dias / 365)
        total   = capital + interes
        ```

        Redondeo a 2 decimales, ROUND_HALF_UP. Si se omite `tna`, se toma la de
        referencia de ArgentinaDatos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capital, dias]
              properties:
                capital: { $ref: '#/components/schemas/Monto' }
                dias: { type: integer, minimum: 30, maximum: 365, example: 30 }
                tna: { type: number, nullable: true, example: 58 }
            example: { capital: 500000, dias: 30, tna: 58 }
      responses:
        '200':
          description: Simulación calculada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  capital: { $ref: '#/components/schemas/Monto' }
                  dias: { type: integer }
                  tna: { type: number }
                  interes: { $ref: '#/components/schemas/Monto' }
                  total: { $ref: '#/components/schemas/Monto' }
                  tea: { type: number }
              example: { capital: 500000, dias: 30, tna: 58, interes: 23835.62, total: 523835.62, tea: 76.23 }
        '400': { $ref: '#/components/responses/ErrorValidacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
  /api/plazos-fijos:
    post:
      tags: [Inversiones]
      summary: Constituir un plazo fijo
      description: Debita el capital de la cuenta y lo inmoviliza hasta el vencimiento.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cuenta_id, capital, dias]
              properties:
                cuenta_id: { type: string, format: uuid }
                capital: { $ref: '#/components/schemas/Monto' }
                dias: { type: integer, minimum: 30, maximum: 365 }
                tna: { type: number, nullable: true }
      responses:
        '201':
          description: Plazo fijo constituido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlazoFijo' }
        '400': { $ref: '#/components/responses/ErrorValidacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '422':
          description: Saldo insuficiente.

    get:
      tags: [Inversiones]
      summary: Listar plazos fijos
      parameters:
        - name: estado
          in: query
          schema: { type: string, enum: [vigente, vencido, acreditado, cancelado_anticipado] }
          description: '`vencido` es el filtro que dice qué hay que acreditar hoy.'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Plazos fijos visibles para el usuario.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginado'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/PlazoFijo' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
...
  /api/plazos-fijos/marcar-vencidos:
    post:
      tags: [Inversiones]
      summary: Marcar los plazos fijos que llegaron a su vencimiento
      description: |
        Pasa a `vencido` los plazos fijos vigentes cuya fecha ya llegó. Es lo que
        hace que el filtro `?estado=vencido` diga qué hay que acreditar hoy.

        Igual que el barrido de mora: se dispara a mano o por cron. Sólo roles
        internos.
      responses:
        '200':
          description: Cantidad de plazos fijos marcados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  vencidos: { type: integer }
              example: { vencidos: 3 }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
  /api/plazos-fijos/{plazoFijoId}/acreditacion:
    post:
      tags: [Inversiones]
      summary: Acreditar al vencimiento o rescatar antes
      description: |
        Cierra el plazo fijo y acredita el dinero en la cuenta.

        | Caso | Interés reconocido |
        |---|---|
        | Al vencimiento | El pactado: `total` |
        | Rescate anticipado | Sólo el de los días transcurridos, a tasa de caja de ahorro |
      parameters:
        - name: plazoFijoId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                anticipada: { type: boolean, default: false }
      responses:
        '200':
          description: Plazo fijo cerrado y acreditado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  plazo_fijo: { $ref: '#/components/schemas/PlazoFijo' }
                  total_acreditado: { $ref: '#/components/schemas/Monto' }
                  dias_transcurridos: { type: integer, example: 30 }
        '400':
          description: Todavía no venció y no se pidió rescate anticipado.
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409':
          description: El plazo fijo ya estaba cerrado.

  # ───────────────────────────── Catálogos ─────────────────────────────
...